Adding `firebase login:ci` command that always re-authenticates and creates a new session token without persisting it
